Improved logic

+++ src/platformAccessory.ts
@@ -64,14 +64,14 @@ export class YeelightBulbPlatformAccessory {
       client.write(payload);
     });
     client.on('data', (data) => {
-      this.platform.log.info('Received: ' + data);
+      this.platform.log.info('Set status - Received: ' + data);
       client.destroy(); // kill client after server's response
     });
     client.on('close', () => {
-      this.platform.log.info('Connection closed');
+      this.platform.log.info('Set status - connection closed');
     });
     setTimeout(() => {
-      if (!client.destroyed){
+      if (!client.destroyed) {
         client.destroy();
       }
     }, 5000);
@@ -96,30 +96,42 @@ export class YeelightBulbPlatformAccessory {
     const port = this.accessory.context.device.location.split('//')[1].split(':')[1];
     const payload = '{"id":1,"method":"get_prop","params":["power"]}\x0D\x0A';
     //send tcp request to yeelight
-    const isOn = await new Promise<boolean>((resolve, reject) => {
-      let isOn = false;
-      const client = new net.Socket();
-      client.connect(port, ip, () => {
-        this.platform.log.info('Connected to Yeelight Bulb, sending payload: ' + payload);
-        client.write(payload);
+    try {
+      const isOn = await new Promise<boolean>((resolve, reject) => {
+        let isOn = false;
+        const client = new net.Socket();
+        client.connect(port, ip, () => {
+          this.platform.log.info('Connected to Yeelight Bulb, sending payload: ' + payload);
+          client.write(payload);
+        });
+        //get response from yeelight
+        client.on('data', (data) => {
+          this.platform.log.info('Check Status - Received: ' + data);
+          const response = JSON.parse(data.toString());
+          if (response.result) {
+            isOn = response.result[0] === 'on';
+          }
+          resolve(isOn);
+          this.platform.log.info('Yeelight Bulb is ' + isOn ? 'on' : 'off');
+          client.destroy(); // kill client after server's response
+        });
+        client.on('close', () => {
+          this.platform.log.info('Check status - connection closed');
+        });
+        //wait for response
+        setTimeout(() => {
+          if (!client.destroyed) {
+            client.destroy();
+          }
+          resolve(isOn);
+        }, 5000);
       });
-      //get response from yeelight
-      client.on('data', (data) => {
-        this.platform.log.info('Received: ' + data);
-        const response = JSON.parse(data.toString());
-        if (response.result) {
-          isOn = response.result[0] === 'on';
-        }
-        resolve(isOn);
-      });
-      //wait for response
-      setTimeout(() => {
-        reject();
-      }, 5000);
-      client.destroy(); // kill client after server's response
-    });
+      return isOn;
+    } catch (error) {
+      this.platform.log.error('Error getting Yeelight Bulb state: ' + error);
+    }
 
-    return isOn;
+    return false;
   }
